refactor: sample app select auth method
