+++ sample/Assets/Scenes/SelectAuthMethod.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bb0668e0c95b745ce8e2f127d5940ede
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
