+++ sample/Assets/Scripts/SampleAppManager.cs
@@ -1,6 +1,5 @@
 public static class SampleAppManager
 {
-    public static bool InitialisedPassport { get; set; }
     public static bool IsConnected { get; set; }
     public static bool UsePKCE { get; set; }
 }
