+++ sample/Assets/Scripts/SelectAuthMethodScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa62d59c085d84ece84de7c1801a9ac1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
